+++ plugins/flow-next/tests/test_flowctl_surface.py
@@ -110,6 +110,7 @@ next
 pilot-log append
 prime classify
 pr-cognitive-aid current
+pr-cognitive-aid html-input
 pr-cognitive-aid render
 pr-cognitive-aid validate
 pr-cognitive-aid write